+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -10,6 +10,12 @@
   AppContext switch to `false` (its default value is `true`).
   ([#3993](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3993))
 
+* Updated OpenTelemetry core component version(s) to `1.15.1`.
+  ([#4020](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4020))
+
+* Fix `http.route` for the root path for Razor Pages.
+  ([#4016](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4016))
+
 ## 1.15.1
 
 Released 2026-Mar-11